Balancing Tree on Insertion
Added checkBalance function to make sure the binary search tree is also an AVL tree. It was implemented in the insert function. main.c was also updated to include an example of a tree which would, normally, be unbalanced, but is fixed by the checkBalance function, with the appropriate insertion function. BF strategy was replaced.